[DSPDC-1864] Retry diff-bq-table on error (#47)
The constituent steps in this workflow often fail due to 502s and other errors from BigQuery. This workflow diffs a staged table against the contents of a Jade table and emits the results to a bucket. In the event of a failure, we should rerun these steps as the new outputs will overwrite any existing outputs (i.e., we are protected against having duplicate entries).

This PR uses an existing argo retry mechanism we are already using for GCS steps and adds it to the overall diff-bq-table template DAG.